feat(docs): add auto height adjustment to iframe builder
